4 link simulation working with hardware

+++ Arduino_Code/Controller_V1/Controller_V1.ino
@@ -1,5 +1,6 @@
 #include <ros.h>
-#include <std_msgs/Float64.h>
+//#include <std_msgs/Float64.h>
+#include <std_msgs/Int16.h>
 
 ////////////////////////////////
 // Pin Definitions
@@ -20,13 +21,13 @@ ros::NodeHandle  nh;
 //TODO: Make custom message type with all four encoder values
 //TODO: Switch to type INT
 
-std_msgs::Float64 encdr_1;
+std_msgs::Int16 encdr_1;
 ros::Publisher pub_encdr_1("/Arduino/01/encoder1/val", &encdr_1);
-std_msgs::Float64 encdr_2;
+std_msgs::Int16 encdr_2;
 ros::Publisher pub_encdr_2("/Arduino/01/encoder2/val", &encdr_2);
-std_msgs::Float64 encdr_3;
+std_msgs::Int16 encdr_3;
 ros::Publisher pub_encdr_3("/Arduino/01/encoder3/val", &encdr_3);
-std_msgs::Float64 encdr_4;
+std_msgs::Int16 encdr_4;
 ros::Publisher pub_encdr_4("/Arduino/01/encoder4/val", &encdr_4);
 
 ////////////////////////////////
